+++ pkg/machinepool/mutate_machinepool_create.go
@@ -8,7 +8,6 @@ import (
 	"k8s.io/api/admission/v1beta1"
 	capiv1alpha3 "sigs.k8s.io/cluster-api/api/v1alpha3"
 
-	"github.com/giantswarm/azure-admission-controller/pkg/generic"
 	"github.com/giantswarm/azure-admission-controller/pkg/mutator"
 )
 
@@ -43,14 +42,6 @@ func (m *CreateMutator) Mutate(ctx context.Context, request *v1beta1.AdmissionRe
 	machinePoolCR := &capiv1alpha3.Cluster{}
 	if _, _, err := mutator.Deserializer.Decode(request.Object.Raw, nil, machinePoolCR); err != nil {
 		return []mutator.PatchOperation{}, microerror.Maskf(parsingFailedError, "unable to parse MachinePool CR: %v", err)
-	}
-
-	patch, err := generic.EnsureOrganizationLabelNormalized(ctx, machinePoolCR)
-	if err != nil {
-		return []mutator.PatchOperation{}, microerror.Mask(err)
-	}
-	if patch != nil {
-		result = append(result, *patch)
 	}
 
 	return result, nil