+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Removed
+
+- Removed Organization label value normalization on CR creation.
+
 ## [1.11.0] - 2020-10-23
 
 ### Added